Calendar Module Update
now working as expected except that when a calendar is assigned to a
syllabus, a copy is made and that copy is assigned to the syllabus. then
when loading from existing calendars, both the assigned calendar and its
original copy are visible. I want the original to become invisible while
its copy is assigned to the syllabus. I haven't figured out a way to do
that